More efficient Watts-Strogatz graphs.
1. Name change:
    smallworld -> watts_strogatz_graph

2. Much faster implementation of WS by minimizing loops.

3. Slight difference with original: First all the nodes that
have to be removed are removed, then new ones are added.
This guarantees that with p=1 there's the same probability
for any two edges to be connected.
In the original model this probability is slightly less for
close neighbors at "the end" of the graph than for other pairs
due to the procedural iteration described in the paper.

4. Added watts_strogatz_digraph, a directed version of the
original that has constant fixed outdegree for all the agents.

    watts_strogatz_digraph(N, 2*k, 1.0) = fixed_outdegree_graph(N,k)

(which is not true for the original model)